update to neo4j 2.0.1

+++ spring-data-neo4j/src/main/java/org/springframework/data/neo4j/support/node/Neo4jHelper.java
@@ -20,6 +20,8 @@ import org.neo4j.graphdb.*;
 import org.neo4j.graphdb.index.Index;
 import org.neo4j.graphdb.index.IndexManager;
 import org.neo4j.graphdb.index.RelationshipIndex;
+import org.neo4j.kernel.GraphDatabaseAPI;
+import org.neo4j.kernel.impl.core.NodeManager;
 import org.neo4j.tooling.GlobalGraphOperations;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -80,10 +82,13 @@ public abstract class Neo4jHelper {
     }
 
     private static void removeNodes(GraphDatabaseService graphDatabaseService, boolean includeReferenceNode) {
+        GraphDatabaseAPI api = (GraphDatabaseAPI) graphDatabaseService;
+        NodeManager nodeManager = api.getDependencyResolver().resolveDependency(NodeManager.class);
         final GlobalGraphOperations globalGraphOperations = GlobalGraphOperations.at(graphDatabaseService);
         for (Node node : globalGraphOperations.getAllNodes()) {
             for (Relationship rel : node.getRelationships(Direction.OUTGOING)) {
                 try {
+                    if (nodeManager.isDeleted(rel)) continue;
                     rel.delete();
                 } catch(IllegalStateException ise) {
                     if (!ise.getMessage().contains("since it has already been deleted")) throw ise;
@@ -93,6 +98,7 @@ public abstract class Neo4jHelper {
         }
         for (Node node : globalGraphOperations.getAllNodes()) {
             try {
+                if (nodeManager.isDeleted(node)) continue;
                 node.delete();
             } catch(IllegalStateException ise) {
                 if (!ise.getMessage().contains("since it has already been deleted")) throw ise;
